Struct eui48::MacAddress

pub struct MacAddress { /* fields omitted */ }

A MAC address (EUI-48)

Methods

impl MacAddress

pub fn new(eui: Eui48) -> MacAddress

Create a new MacAddress from [u8; 6].

pub fn from_bytes(bytes: &[u8]) -> Result<Self, ()>

Create a new MacAddress from a byte slice.

Returns an error (without any description) if the slice doesn't have the proper length.

pub fn nil() -> MacAddress

Returns empty EUI-48 address

pub fn broadcast() -> MacAddress

Returns 'ff:ff:ff:ff:ff:ff', a MAC broadcast address

pub fn is_nil(&self) -> bool

Returns true if the address is '00:00:00:00:00:00'

pub fn is_broadcast(&self) -> bool

Returns true if the address is 'ff:ff:ff:ff:ff:ff'

pub fn is_unicast(&self) -> bool

Returns true if bit 1 of Y is 1 in address 'xY:xx:xx:xx:xx:xx'

pub fn is_multicast(&self) -> bool

Returns true if bit 1 of Y is 1 in address 'xY:xx:xx:xx:xx:xx'

pub fn is_universal(&self) -> bool

Returns true if bit 2 of Y is 0 in address 'xY:xx:xx:xx:xx:xx'

pub fn is_local(&self) -> bool

Returns true if bit 2 of Y is 1 in address 'xY:xx:xx:xx:xx:xx'

pub fn to_canonical(&self) -> String

Returns a String representation in the format '00-00-00-00-00-00'

pub fn to_hex_string(&self) -> String

Returns a String representation in the format '00:00:00:00:00:00'

pub fn to_dot_string(&self) -> String

Returns a String representation in the format '0000.0000.0000'

pub fn to_hexadecimal(&self) -> String

Returns a String representation in the format '0x000000000000'

pub fn to_string(&self, fmt: MacAddressFormat) -> String

Returns a String in the format selected by fmt

pub fn parse_str(s: &str) -> Result<MacAddress, ParseError>

Parses a String representation from any format supported

pub fn as_bytes<'a>(&'a self) -> &'a [u8]

Return the internal structure as a slice of bytes